[refactor/#75] - id 타입 Int에서 Long으로 변경

+++ data/src/main/java/com/umc/data/api/AttendanceApi.kt
@@ -20,16 +20,16 @@ interface AttendanceApi {
 
     @GET(Endpoints.Attendance.PENDING)
     suspend fun getPendingUsers(
-        @Path("scheduleId") scheduleId: Int
+        @Path("scheduleId") scheduleId: Long
     ): ApiResponse<List<AdminPendingUserResponse>>
 
     @POST(Endpoints.Attendance.APPROVE)
     suspend fun approveAttendance(
-        @Path("recordId") recordId: Int
+        @Path("recordId") recordId: Long
     ): ApiResponse<Unit>
 
     @POST(Endpoints.Attendance.REJECT)
     suspend fun rejectAttendance(
-        @Path("recordId") recordId: Int
+        @Path("recordId") recordId: Long
     ): ApiResponse<Unit>
 }
