+++ domain/src/main/java/com/umc/domain/repository/attendance/AttendanceRepository.kt
@@ -8,7 +8,7 @@ import com.umc.domain.model.request.attendance.AttendanceCheckRequest
 interface AttendanceRepository {
     suspend fun getAttendanceAvailable(): ApiState<List<UserCheckAvailable>>
     suspend fun postAttendanceCheck(request: AttendanceCheckRequest): ApiState<String>
-    suspend fun getPendingUsers(scheduleId: Int): ApiState<List<AdminPendingUser>>
-    suspend fun approveAttendance(recordId: Int): ApiState<Unit>
-    suspend fun rejectAttendance(recordId: Int): ApiState<Unit>
+    suspend fun getPendingUsers(scheduleId: Long): ApiState<List<AdminPendingUser>>
+    suspend fun approveAttendance(recordId: Long): ApiState<Unit>
+    suspend fun rejectAttendance(recordId: Long): ApiState<Unit>
 }
